moved testTrackingSubmoduleBranches to GItAPITestUpdateCliGit
